"Move DemoRegistrationRequest to 'dto.demo' package for improved structure."
